vertice-terminal/requirements.txt
typer>=0.12.3
rich>=13.0.0
httpx>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0
questionary>=2.0.0
diskcache>=5.6.0
typing-extensions>=4.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
keyring>=24.0.0
cryptography>=41.0.0
google-generativeai>=0.3.0
prompt-toolkit>=3.0.0
tomli>=2.0.0
tomli-w>=1.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-httpx>=0.29.0
# Linter and formatting tools
black>=24.0.0
flake8>=7.0.0
mypy>=1.0.0
bandit>=1.7.0
# Type stubs
types-PyYAML>=6.0.0
